% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRGboxplot.R
\name{DRGboxplot}
\alias{DRGboxplot}
\title{DRGboxplot: Display a boxplot of payments by DRG code}
\usage{
DRGboxplot(df, vary)
}
\arguments{
\item{df}{some data frame. Here, we choose DRG.}

\item{vary}{the input for some payment. Here, we pick one from the average
Medicare payments, the average total payment, and the average covered charges.}
}
\value{
The boxplot of payments by DRG code
}
\description{
DRGboxplot: Display a boxplot of payments by DRG code
}
\examples{
DRGboxplot(DRG, DRG$Average.Total.Payments)

}
