% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRGstatistics.R
\name{DRGstatistics}
\alias{DRGstatistics}
\title{DRGstatistics: Calculate statistics over all of the DRG codes for average Medicare payments}
\usage{
DRGstatistics(df, option)
}
\arguments{
\item{df}{the input of some data frame. Here, we choose DRG.}

\item{option}{the input of either 'mean', 'sd', or 'median'.}
}
\value{
Displays the mean, sd, or median corresponding to DRG codes
}
\description{
DRGstatistics: Calculate statistics over all of the DRG codes for average Medicare payments
}
\examples{
DRGstatistics(DRG, 'mean')

}
